conclusão do dash
