+++ Mech-Dungeon/Assets/Scripts/Habilidades/CadaCarta.cs
@@ -1,0 +1,12 @@
+using System.Collections;
+using System.Collections.Generic;
+using UnityEngine;
+
+public class CadaCarta : ScriptableObject
+{
+    public Sprite sprite;
+    public string Nome;
+    public string Descricao;
+    public int QualidadeDeManufatura;
+    public int AtributoEspecial;
+}
