+++ Mech-Dungeon/Assets/Scripts/Habilidades/CadaCarta.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: c8d3fd488d54f5740a1ea0f0abf8a885
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
